src/backend/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
psycopg2-binary==2.9.9      # Postgres 드라이버
faiss-cpu==1.8.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
numpy==1.26.4
python-dotenv==1.0.1        # .env 파일 로드
google-generativeai==0.8.3  # Phase 5: Gemma API fallback
